requirements.txt
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0
pandas==1.5.3